refactor keyboard again use uint16 like mouse to ensure that keycodes can be used for both classses

+++ Dsr/src/Events/Application/InputEvents/KeyboardEvents.cpp
@@ -6,12 +6,12 @@ namespace dsr
 {
 	namespace events
 	{
-		KeyDownEvent::KeyDownEvent(const uint8_t& keyCode)
+		KeyDownEvent::KeyDownEvent(const uint16_t& keyCode)
 			: m_keyCode(keyCode)
 		{
 		}
 
-		KeyUpEvent::KeyUpEvent(const uint8_t& keyCode)
+		KeyUpEvent::KeyUpEvent(const uint16_t& keyCode)
 			: m_keyCode(keyCode)
 		{
 		}
